refactor: UserContext 정의
