+++ src/main/java/followMe/hub_server/stock/application/service/UserContext.java
@@ -1,0 +1,5 @@
+package followMe.hub_server.stock.application.service;
+
+import java.util.UUID;
+
+public record UserContext(UUID userId, UserRole role, UUID hubId, UUID vendorId) {}
