+++ src/main/java/followMe/hub_server/stock/application/service/UserRole.java
@@ -1,4 +1,4 @@
-package followMe.hub_server.stock.domain;
+package followMe.hub_server.stock.application.service;
 
 public enum UserRole {
   MASTER,
